requirements.txt
langchain
langchain-openai
langchain-community
faiss-cpu
streamlit
pandas
python-dotenv
tiktoken
pydantic
requests
